feat: add new database property types including Button, Verification, UniqueId, Url, Date, Email, Title, and Files
